use "sourceLibrary"; add 'hostname' to betterC API
This library has different constructors (and unfortunately slightly
different ABI due to dlang issue 22031) depending on if it's compiled
with -betterC, but if dub has a feature to propaganda -betterC flags
to dependencies, it's not clearly documented. By using a targetType of
"sourceLibrary", this library gets compiled into the same compiler call
as its users, so inherits their -betterC flag.

It was just an error to omit 'hostname' from the betterC API.